Test_case_generation_service/refactored/requirements.txt
google-cloud-aiplatform
beautifulsoup4
lxml
selenium-stealth
webdriver-manager
flask 
flask-cors
google-cloud-firestore 
google-cloud-storage 
requests 
gunicorn
typing_extensions >= 3.7.3.4
GitDB
GitPython
